Bottom sheet "description of photo of day" moved to page fragment.
