+++ app/src/main/res/layout/fragment_main.xml
@@ -22,10 +22,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"/>
 
-        <include
-            android:id="@+id/including_bottom_sheet"
-            layout="@layout/bottom_sheet_layout" />
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </FrameLayout>
